>> load xml config
change:
	1. split build.c add ergodic call back ergodic_item/ergodic_group
	2. read/write xml config :
		rwxml.c
	3. output Minishell structure "struct cmd_prompt" :
		wcfile.c
	4. this tool base structure operation
		build-structure.c
	5. clear most compile Warning

clear most compile WARNING
