+++ def.xml
@@ -1,15 +1,22 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<minishell>
-  <group name="boot1">
-    <item name="cmd11222" alias="" help=""></item>
-    <item name="cmd12" alias="" help=""></item>
-  </group>
-  <group name="boot2">
-    <item name="cmd21" alias="" help="help 2"></item>
-    <item name="cmd22" alias="" help="help 3"></item>
-  </group>
-  <group name="boot3">
-    <item name="cmd31" alias="c31" help=""></item>
-    <item name="cmd32" alias="c32" help=""></item>
-  </group>
-</minishell>
+<?xml version="1.0" encoding="ISO-8859-1"?>
+
+<bookstore>
+
+<group name="boot1" abc="sd">
+<item name="cmd11222" boot="boot2">ss12s</item>
+<item name="cmd12" boot="boot3" >ss12s</item>
+</group>
+
+<group name="boot2">
+<item name="cmd21"  boot="boot1" help="help 2">sss33</item>
+<item name="cmd22"  boot=""  help="help 3">sss33</item>
+</group>
+
+<group name="boot3">
+<item  name="cmd31" boot="" alias="c31">sss33</item>
+<item name="cmd32"  boot="" alias="c32">sss33</item>
+</group>
+
+
+
+</bookstore>
