+++ script/allprj.mk
@@ -23,3 +23,10 @@ alldisclean:
 
 	$(MAKE) DP=pi4 disclean --no-print-directory
 	$(MAKE) DP=pi5 disclean --no-print-directory
+
+# strip all output file STRTAB section
+.PHONY:allstrip
+allstrip:
+	$(MAKE) DP=pi1 strip --no-print-directory
+	$(MAKE) DP=pi4 strip --no-print-directory
+	$(MAKE) DP=pi5 strip --no-print-directory
