+++ script/listprj.mk
@@ -5,11 +5,11 @@
 # pixxx=script/default
 # pixxx_arg=[all,mlib,elf...] one of us, [all, mlib,elf...] is top path Makefile operation
 	pi1=script/minishellex_example
-		pi1_arg=bin
+		pi1_arg=elf
 	pi4=script/tool-build
-		pi4_arg=bin
+		pi4_arg=elf
 	pi5=script/test
-		pi5_arg=bin
+		pi5_arg=elf
 
 	pi2=script/prj_liba
 		pi2_arg=mlib
